CORE-170: shared: add seqOfN codec
