+++ shared/src/main/scala/coop/rchain/scodec/codecs/SeqCodec.scala
@@ -1,0 +1,21 @@
+package coop.rchain.scodec.codecs
+
+import scodec.bits.BitVector
+import scodec.{Attempt, Codec, DecodeResult, Decoder, Encoder, SizeBound}
+
+import scala.collection.immutable.Seq
+
+final class SeqCodec[A](codec: Codec[A], limit: Option[Int] = None) extends Codec[Seq[A]] {
+
+  def sizeBound: SizeBound = limit match {
+    case None      => SizeBound.unknown
+    case Some(lim) => codec.sizeBound * lim.toLong
+  }
+
+  def encode(list: Seq[A]): Attempt[BitVector] = Encoder.encodeSeq(codec)(list)
+
+  def decode(buffer: BitVector): Attempt[DecodeResult[Seq[A]]] =
+    Decoder.decodeCollect[Seq, A](codec, limit)(buffer)
+
+  override def toString = s"seq($codec)"
+}
